use uintptr_t type to save pointer value

+++ src/osgPlugins/osgjs/WriteVisitor.cpp
@@ -225,7 +225,7 @@ JSONObject* createImage(osg::Image* image, bool inlineImages, int maxTextureDime
             // no image file so use this inline name image and create a file
             std::stringstream ss;
             ss << osgDB::getFilePath(baseName) << osgDB::getNativePathSeparator();
-            ss << (long int)image << ".inline_conv_generated.png"; // write the pointer location
+            ss << (uintptr_t)image << ".inline_conv_generated.png"; // write the pointer location
             std::string filename = ss.str();
             if (osgDB::writeImageFile(*image, filename)) {
                 image->setFileName(filename);
